RFT:move visualization functions to the visualization subpackage
